Moved the parser into its own file
